+++ src/pages/pedidos/utils.jsx
@@ -1,0 +1,7 @@
+export function formatearPrecio(valor) {
+  return valor.toLocaleString('es-CO', {
+    style: 'currency',
+    currency: 'COP',
+    maximumFractionDigits: 0
+  });
+}
